requirements.txt
Flask==2.1.0
plaid_python==9.1.0
python-dotenv==0.15.0
itsdangerous==2.0.1
pandas==1.4.2
pymongo==4.3.3